requirements.txt
streamlit
pydantic
pydantic_settings
google-generativeai
pymongo
pyyaml
graphviz